Extract CreateLocation() as extension method
