+++ analyzers/src/SonarAnalyzer.Common/Rules/Hotspots/DisablingRequestValidationBase.cs
@@ -110,7 +110,7 @@ namespace SonarAnalyzer.Rules
             foreach (var pages in doc.XPathSelectElements("configuration/system.web/pages"))
             {
                 if (pages.GetAttributeIfBoolValueIs("validateRequest", false) is { } validateRequest
-                    && CreateLocation(webConfigPath, validateRequest) is { } location)
+                    && validateRequest.CreateLocation(webConfigPath) is { } location)
                 {
                     c.ReportDiagnosticWhenActive(Diagnostic.Create(rule, location));
                 }
@@ -124,26 +124,11 @@ namespace SonarAnalyzer.Rules
                 if (httpRuntime.Attribute("requestValidationMode") is { } requestValidationMode
                     && decimal.TryParse(requestValidationMode.Value, NumberStyles.Number, CultureInfo.InvariantCulture, out var value)
                     && value < MinimumAcceptedRequestValidationModeValue
-                    && CreateLocation(webConfigPath, requestValidationMode) is { } location)
+                    && requestValidationMode.CreateLocation(webConfigPath) is { } location)
                 {
                     c.ReportDiagnosticWhenActive(Diagnostic.Create(rule, location));
                 }
             }
-        }
-
-        private static Location CreateLocation(string path, XAttribute attribute)
-        {
-            // IXmlLineInfo is 1-based, whereas Roslyn is zero-based
-            var startPos = (IXmlLineInfo)attribute;
-            if (startPos.HasLineInfo())
-            {
-                // LoadOptions.PreserveWhitespace doesn't preserve whitespace inside nodes and attributes => there's no easy way to find full length of a XAttribute.
-                var length = attribute.Name.ToString().Length;
-                var start = new LinePosition(startPos.LineNumber - 1, startPos.LinePosition - 1);
-                var end = new LinePosition(startPos.LineNumber - 1, startPos.LinePosition - 1 + length);
-                return Location.Create(path, new TextSpan(start.Line, length), new LinePositionSpan(start, end));
-            }
-            return null;
         }
     }
 }
